Create Arithmetic with array.asm
Input 5 numbers, add the first 3 numbers and subtract the last 2 numbers from it. The solution is a generalized one. Then number of items to be added subtracted can be changed in the code easily. The straightforward answer to the question is more efficient in terms of code, but is less update friendly.